item crud completed, update and delete fixed

+++ controllers/inventoryController.js
@@ -1,5 +1,6 @@
 const {Item} = require('../models/item')
 const _ = require('lodash')
+const itemModel = require('../models/item')
 
 exports.getItems = async(req, res) => {
     try {
@@ -24,14 +25,15 @@ exports.createItem = async (req, res) => {
     }
 }
 
-exports.getItem = async(req, res) => {
-    res.send(req.url)
-}
-
 exports.renderUpdateItem = async(req, res) => {
     try {
-        const item = await Item.findById(req.params.id)
-        res.render('inventory/edit', {item})
+        const isFound = await findItem(req.params.id)
+        if(!isFound){
+            req.flash('error', 'Item not found')
+            return res.redirect('/inventory')
+        }
+        // const item = await Item.findById(req.params.id)
+        res.render('inventory/edit', {item: isFound})
     } catch (error) {
         req.flash('error', error.message)
         res.redirect('/inventory')
@@ -39,7 +41,18 @@ exports.renderUpdateItem = async(req, res) => {
 }
 exports.updateItem = async(req, res) => {
     try {
-        await Item.findByIdAndUpdate(req.params.id, _.pick(req.body, ['price', 'stock', 'manufacturer', 'manufacturer_price']), {runValidators: true})
+        // validate request data
+        const {error} = itemModel.ValidateUpdate.validate(req.body)
+        if(error){
+            req.flash('error', error.message)
+            return res.redirect(`/inventory/${req.params.id}/edit`)
+        }
+        const isFound = await findItem(req.params.id)
+        if(!isFound){
+            req.flash('error', 'Item not found')
+            return res.redirect('/inventory')
+        }
+        await Item.findByIdAndUpdate(req.params.id, {..._.pick(req.body, ['price', 'stock', 'manufacturer', 'manufacturer_price']), modified_date: new Date}, {runValidators: true})
         req.flash('success', 'Item updated')
         res.redirect(`/inventory/${req.params.id}/edit`)
     } catch (error) {
@@ -50,11 +63,29 @@ exports.updateItem = async(req, res) => {
 
 exports.deleteItem = async(req, res) => {
     try {
+        const isFound = await findItem(req.params.id)
+        if(!isFound){
+            req.flash('error', 'Item not found')
+            return res.redirect('/inventory')
+        }
         await Item.findByIdAndDelete(req.params.id)
         req.flash('success', 'Item deleted')
         res.redirect('/inventory')
     } catch (error) {
         req.flash('error', error.message)
         res.redirect('/inventory')
+    }
+}
+
+// impure
+async function findItem(id){
+    try {
+        const data = await Item.findById(id)
+        if(!data){
+            return false
+        }
+        return data
+    } catch (error) {
+        console.log(error.message)
     }
 }
